Refactor: centralize GPU context management in GpuContextBase
Move ILGPU context/accelerator logic to new GpuContextBase class, reducing duplication across extension classes. All GPU-bound extensions now use GpuContextBase for context access and disposal. Add support for both default and fast math GPU contexts. Update correlation test to use float arrays for large datasets.

+++ Mavusi.Linq.DataScience.Tests/GpuBound/CorrelationExtensionsTests.cs
@@ -143,8 +143,8 @@ public class CorrelationExtensionsTests
     {
         // Arrange - large dataset to benefit from GPU parallelization
         var size = 10000;
-        var x = Enumerable.Range(0, size).Select(i => (double)i).ToArray();
-        var y = Enumerable.Range(0, size).Select(i => (double)i * 2 + 5).ToArray();
+        var x = Enumerable.Range(0, size).Select(i => (float)i).ToArray();
+        var y = Enumerable.Range(0, size).Select(i => (float)i * 2 + 5).ToArray();
 
         // Act
         var result = x.CorrelationGpu(y);
